backend/requirements.txt
fastapi
uvicorn
python-multipart      # for handling file uploads
passlib[bcrypt]
PyMySQL      # MySQL driver
SQLAlchemy      # ORM
python-jose[cryptography]     # for JWT authentication
pydantic